libports: avoid implicit conversions
Issue #23

+++ repos/libports/src/lib/lwip/platform/rand.cc
@@ -24,7 +24,7 @@ genode_uint32_t genode_rand()
 	state = oldstate * 6364136223846793005ULL + inc;
 
 	// Calculate output function (XSH RR), uses old state for max ILP
-	uint32_t xorshifted = ((oldstate >> 18u) ^ oldstate) >> 27u;
-	uint32_t rot = oldstate >> 59u;
+	uint32_t xorshifted = (uint32_t)(((oldstate >> 18u) ^ oldstate) >> 27u);
+	uint32_t rot        = (uint32_t)(oldstate >> 59u);
 	return (xorshifted >> rot) | (xorshifted << ((-rot) & 31));
 }
